+++ sdk/core/Azure.Core.Experimental/src/HttpPipelineExtensions.cs
@@ -4,9 +4,8 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using Azure.Core.Pipeline;
 
-namespace Azure.Core.Experimental
+namespace Azure.Core.Pipeline
 {
     /// <summary>
     /// Extensions to HttpPipeline to support RequestOptions.
@@ -19,9 +18,20 @@ namespace Azure.Core.Experimental
         /// <param name="pipeline"></param>
         /// <param name="options">The message options.</param>
         /// <returns>The message.</returns>
-        public static HttpMessage CreateMessage(this HttpPipeline pipeline, RequestOptions options)
+        public static HttpMessage CreateMessage(this HttpPipeline pipeline, RequestOptions? options)
         {
+            // TODO: This method will be added as a method on HttpPipeline directly
+            // when RequestOptions moves to core. At that time, we expect RequestContext
+            // to inherit from RequestOptions, so copying RequestOptions to
+            // RequestContext can be removed.
+
+            if (options == null)
+            {
+                return pipeline.CreateMessage();
+            }
+
             RequestContext context = new RequestContext();
+
             context.ErrorOptions = options.ErrorOptions;
 
             if (options.Policies != null)